Fix building docs for release-5.0.x
The new SCMTools registry work caused a problem trying to build docs.
The issue was that we were using `AppConfig.ready` to populate the
registry, but doing so will do some database accesses (in order to pull
any legacy Tool instances that don't exist in entry points into the
registry). According to the documentation for `ready`, we should avoid
doing any DB queries here, precisely because it can affect running
management commands.

This change moves the registry population back into a handler for the
`initializing` signal, along with a comment explaining why.

While it's apparently not necessary for the docs build, I've also added
the new `scmtool_id` field to the fixtures for the docs DB.

Testing Done:
- Ran unit tests.
- Built docs.

Reviewed at https://reviews.reviewboard.org/r/12370/