Remove double Q(Q(...)) objects when comparing in Profile unit tests.
A fix is in for Djblets to avoid an annoying quirk with Django where a
`filter()` operation can create a nested `Q(Q(...))`. This was annoying
to compare, and Djblets now normalizes that.

This updates the `Profile` unit tests to do the same.

Testing Done:
Unit tests pass.

Reviewed at https://reviews.reviewboard.org/r/12358/