Fix a bad render call when rendering change form pages.
In 2020, a change (0ebfa480a6) was made to `ChangeFormPageView` that
updated a `render()` call to `renderPage()`. This was done as part of a
series of fixes to page rendering, except that this particular update
was made to an `AdminInlineGroupView`. This function doesn't exist
there, so this failed. Apparently this has gone unnoticed, and wasn't
really that big of a breakage.

This change fixes that call, reverting it back to a `render()`. This
will be backported to 4.0.x.

Testing Done:
Loaded the User change page. Verified it no longer errored out with
an invalid function call.

Reviewed at https://reviews.reviewboard.org/r/12387/