Use Console for pre/post-upgrade output.
This integrates our `Console` object (used by `rb-site` and `manage.py`)
into our pre/post-upgrade code. This is used to provide nicer
wrapped/styled output for errors that fits in with the rest of the
upgrade appearance.

There are some issues with the display of the output, but that's an
issue in `Console`, and will be addressed separately.

Testing Done:
Triggered the two SCMTool failure error cases (failure to import, and
conflicts between `Tool` entries). Saw the formatted output.

Reviewed at https://reviews.reviewboard.org/r/12392/