Add a new extension hook for registering SCMTools.
Now that SCMTools are backed by a registry instead of relying entirely
on entry points and the `registerscmtools` management command, we can
make it easier for extensions to add new ones. This change adds a new
extension hook, `SCMToolHook`, for doing so. It works very similarly to
the existing hook for hosting services.

Testing Done:
- Ran unit tests.
- Built and reviewed docs.

Reviewed at https://reviews.reviewboard.org/r/12395/