Fix a regression in displaying diff fragments in comments.
When upgrading to Django 3.2, we converted all of our template tags and
filters to use `mark_safe()` instead of setting `is_safe = True`, with
the exception of `showextrawhitespace`.

This was turning safe HTML into an unsafe string, and the result was
that any comments shown in comments in a review or in e-mails was
turning into big blobs of escaped HTML.

This change fixes it to use `mark_safe()`, and adds some unit tests to
ensure this never regresses.

Testing Done:
Viewed comments in a diff that exhibited this problem. Saw that the
content was correct in all chunks of the diffs I expanded.

Unit tests pass.

Reviewed at https://reviews.reviewboard.org/r/12402/