Allow timestamp to be set for comment models in test cases
The create methods for comments in `reviewboard.testing.TestCase` does not
properly set the timestamp when a `timestamp` argument is passed. This is
because the timestamp is being overriden when when the comment object is saved.
This change allows `reviewboard.testing.TestCase` to properly set the timestamp
when a `timestamp` argument is passed.

Testing Done:
Ran all unit tests in `reviewboard.webapi.tests`, which make use of the create
comment methods.

Reviewed at https://reviews.reviewboard.org/r/12394/