Enable usage of newer psycopg2-binary packages for Postgres.
Now that we're on a modern version of Django, we no longer need to worry
about capping the version of the `psycopg2-binary` dependency for
Postgres. We previously needed to cap it in order to avoid a
compatibility issue with newer versions of the package and older
versions of Django.

This change removes the cap. At this point, any version installed is
considered fine for our needs.

Testing Done:
Installed the package via `ReviewBoard[postgres]` successfully.

Reviewed at https://reviews.reviewboard.org/r/12416/