Update unit tests to change non-unique state between models.
We've recently uncovered a bug that allows certain models
(`LocalSiteProfile`, `Group`, and `Repository`) to have duplicate
entries containing the same "unique" state when not associated with a
`LocalSite`.

The fix for that is still in development, but it has uncovered many unit
tests that introduce duplicate state for these objects.

This change fixes up that state, preparing the test suite for the
upcoming fix.

Testing Done:
Unit tests all pass with and without the pending change to prevent
duplicate state.

Reviewed at https://reviews.reviewboard.org/r/12426/