Bump Pygments dependency to 2.12.
Pygments is generally quite safe to upgrade, and we want to make sure
that everyone is more or less on the same page as far as support and bug
fixes. This change bumps our requirement up to the latest release, 2.12.

Testing Done:
Ran setup.py develop and saw that I got the correct version.