Fixed a typo in the installation docs
There's a missing `install` in the `pip` example for installing in-development
releases. This change fixes the typo for schmucks like me who want to copy and
paste the example and don't realize it's missing the `install`.

Testing Done:
Built the docs and looked at the page.

Reviewed at https://reviews.reviewboard.org/r/12446/