Allow Djblets to patch collections when running rbext.
Most of our command line tools were updated to allow Djblets to patch
the `collections` module on Python 3.10, but `rbext` was not. This meant
that if you're using Python 3.10, you're unable to invoke `rbext` at all
on Python 3.10. Whether this has been hit in the wild is unknown, but
it's blocking a new CI pipeline.

We now `import djblets` early in the module, giving Djblets the
opportunity to apply the patch.

Testing Done:
Successfully ran `rbext` using Python 3.10.

Reviewed at https://reviews.reviewboard.org/r/12449/