Override cache settings during rb-site invocation.
The change to move from `MemcachedCache` to `PymemcacheCache` broke
things when upgrading from a Python 2-based Review Board 3.0.x to Python
3-based 5.0 beta 1. The problem is that various code paths during
`django.start()` ended up triggering cache operations, and trying to set
up the old (pre-migrated) `MemcachedCache` would fail due to a missing
module. We migrate this setting during the upgrade process, but we never
got that far.

This change fixes rb-site to override any stored cache settings and just
use the local-memory cache for the duration of the command. This is
guaranteed to work regardless of what modules happen to be installed.

Testing Done:
Ran rb-site upgrade on a site where the old environment used the
`memcached` package, and the new environment used the `pymemcache`
package.

Reviewed at https://reviews.reviewboard.org/r/12457/