Fix CACHES migration.
When I implemented the migration code to update the CACHES setting from
the old `MemcachedCache` to the new `PymemcacheCache`, I tested the
actual migration implementation, but had just made an assumption about
the code that checks if we need to run it. It turns out this had two
problems--we were skipping one step that needed to check if a settings
update was required, and in the running context of rb-site, the key that
we need to check is `forwarded_backend` instead of `default`.

Testing Done:
Ran a real `rb-site upgrade` and saw that my settings_local.py file was
correctly updated.

Reviewed at https://reviews.reviewboard.org/r/12458/