Fix rendering of the OAuth application client secret widget.
This change mirrors a similar one to
`django.forms.widgets.CopyableInputWidget`. A change in Django's form
rendering meant that calling the TextInput superclass' `render()` method
was still using `self.template_name`, which would end up rendering two
copies of the template, one of which was broken. This change makes it so
we explicitly render the text input template for that part of the
display.

Testing Done:
Loaded the OAuth application edit form. Saw that the client secret
widget displayed correctly, and that the "Regenerate client secret"
button worked correctly.

Reviewed at https://reviews.reviewboard.org/r/12466/