Use CopyableTextInput for install key.
We've had a special form input widget that makes it easy to copy the
contents for a while, but for some reason we never transitioned the
install key over to use it. This change does that.

Testing Done:
Loaded the support settings page and saw that the install key now used
the copyable text input. Clicked the copy icon and saw the install key
in my clipboard.

Reviewed at https://reviews.reviewboard.org/r/12467/